requirements.txt
annotated-types==0.6.0
anyio==4.3.0
archspec @ file:///croot/archspec_1697725767277/work
boltons @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/boltons_1699245321609/work
Brotli @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/brotli-split_1699238731048/work
certifi @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_dfez5lkpj0/croot/certifi_1707229180975/work/certifi
cffi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_44mvycfmuf/croot/cffi_1700254431144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
conda @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_05eqn2x4t4/croot/conda_1708369127465/work
conda-content-trust @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/conda-content-trust_1699284950649/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1702997573971/work/src
conda-package-handling @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_9ak7pw0efb/croot/conda-package-handling_1700254771505/work
conda_package_streaming @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/conda-package-streaming_1699241680711/work
cryptography @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_85yzbl_qze/croot/cryptography_1702070288916/work
Distance==0.1.3
distro @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/distro_1701806088279/work
filelock==3.13.1
fsspec==2024.3.0
g2p-en==2.1.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/idna_1699239456905/work
inflect==7.0.0
joblib==1.3.2
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
libmambapy @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/mamba-split_1701809865327/work/libmambapy
llvmlite==0.42.0
menuinst @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_0bl_d_k5ae/croot/menuinst_1706732940486/work
MouseInfo==0.1.3
nltk==3.8.1
numba==0.59.0
numpy==1.26.4
openai==1.14.1
opencv-python==4.9.0.80
packaging @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/packaging_1699238434731/work
pandas==2.2.1
pillow==10.2.0
platformdirs @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/platformdirs_1701805067573/work
pluggy @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pluggy_1699238591533/work
PyAutoGUI==0.9.54
pycosat @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pycosat_1699246746992/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.6.4
pydantic_core==2.16.3
pyfoal==0.0.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.2
pyobjc-framework-Cocoa==10.2
pyobjc-framework-Quartz==10.2
pypar==0.0.5
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pysocks_1699239289103/work
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_4fp_3tflo8/croot/requests_1707355574897/work
resampy==0.4.3
ruamel.yaml @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/ruamel.yaml_1699247046910/work
rubicon-objc==0.4.7
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
tokenizers==0.15.2
tqdm @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/tqdm_1707340842215/work
truststore @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/truststore_1701811390284/work
typing_extensions==4.10.0
tzdata==2024.1
urllib3 @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_cb4f5apn6b/croot/urllib3_1707770574455/work
wheel==0.41.2
zstandard @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/zstandard_1699241513306/work
